fix pauses
