--- Assets/Scripts/CompositePause.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aa6af5d8f5992484cbf7a0c1eaad6193
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
