+++ Assets/Scripts/MenuUI/uiManager/SwitchWindow.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 public class SwitchWindow : MonoBehaviour
 {
@@ -6,10 +7,13 @@ public class SwitchWindow : MonoBehaviour
     [SerializeField] private GameObject menuUI;
     [SerializeField] private FollowFocusedButton followFocusedButton;
     [SerializeField] private Pause volumeControllerPause;
+    private Pause playerPause;
 
     void Start()
     {
         _inputSetting = InputSetting.Load();
+        DebugLogger.Log("OnConversationStart/End called.");
+        SceneManager.sceneLoaded += SceneLoaded;
     }
     void Update()
     {
@@ -20,22 +24,28 @@ public class SwitchWindow : MonoBehaviour
             {
                 SetWindowActive(false);
                 volumeControllerPause.UnPauseAll();
-                MenuUIManager.Instance.PlayerPause.UnPauseAll();
+                playerPause.UnPauseAll();
             }
             else
             {
                 SetWindowActive(true);
                 volumeControllerPause.PauseAll();
-                MenuUIManager.Instance.PlayerPause.PauseAll();
+                playerPause.PauseAll();
             }
         }
         else if (_inputSetting.GetCancelKeyDown() && menuUI.activeInHierarchy)
         {
             SetWindowActive(false);
-            MenuUIManager.Instance.PlayerPause.UnPauseAll();
+            playerPause.UnPauseAll();
         }
     }
-
+    
+    void SceneLoaded(Scene nextScene, LoadSceneMode mode)
+    {
+        if (!nextScene.name.Contains("room")) return;
+        
+        playerPause = GameObject.Find("Pause").GetComponent<Pause>();
+    }
 
     private void SetWindowActive(bool isActive)
     {
