+++ Assets/Scripts/SoundManager/VolumeControlManager.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using UnityEngine.Serialization;
 using UnityEngine.UI;
 
 public class VolumeControlManager : MonoBehaviour
@@ -8,14 +9,14 @@ public class VolumeControlManager : MonoBehaviour
     [SerializeField] private GameObject volumeControlObject;
     private Pause playerPause;
     [SerializeField] private Pause menuUIPause;
-    [SerializeField] private Pause soundManagerPause;
+    [SerializeField] private Pause pause;
 
     void Start()
     {
         SceneManager.sceneLoaded += SceneLoaded;
         playerPause ??= FindPause();
-        ConversationTextManager.Instance.OnConversationStart += soundManagerPause.PauseAll;
-        ConversationTextManager.Instance.OnConversationEnd += soundManagerPause.UnPauseAll;
+        ConversationTextManager.Instance.OnConversationStart += pause.PauseAll;
+        ConversationTextManager.Instance.OnConversationEnd += pause.UnPauseAll;
     }
 
     void SceneLoaded(Scene nextScene, LoadSceneMode mode)
